fix(d-routes): added routes file
