+++ src/components/index.js
@@ -1,3 +1,2 @@
 export * from './Layout'
-export * from './Nav'
 export * from './SideBar'
