--- src/routes/Navigate.jsx
@@ -1,26 +1,0 @@
-import App from 'App'
-import { Login, Home } from 'pages'
-import { createBrowserRouter } from 'react-router-dom'
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <App />,
-    children: [
-      // {
-      //   path: '.',
-      //   element: <Home />
-      // }
-    ]
-  },
-  {
-    path: '/login',
-    element: <Login />
-  },
-  {
-    path: '/register',
-    element: <Login />
-  }
-])
-
-export default router
